third_party/WebKit/LayoutTests/compositing/will-change/parse-will-change-expected.txt

Test that setting and getting will-change works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting will-change set through CSS
PASS getComputedStyle(willChangeOpacity, '').getPropertyValue('will-change') is 'opacity'
PASS getComputedStyle(willChangeScrollPosition, '').getPropertyValue('will-change') is 'scroll-position'
PASS getComputedStyle(willChangeContents, '').getPropertyValue('will-change') is 'contents'
PASS getComputedStyle(willChangeMultiple, '').getPropertyValue('will-change') is 'contents, opacity, left, -webkit-transform'
PASS getComputedStyle(willChangeWithArbitraryIdent, '').getPropertyValue('will-change') is 'opacity, top'

Test initial value of will-change
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'

Test getting and setting will-change through JS
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'opacity'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'scroll-position'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'contents'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'contents, scroll-position, opacity, -webkit-transform'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'opacity, top'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'opacity, top'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'opacity, top'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'

Test that illegal will-change values are disallowed
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'

Test the value 'initial'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'opacity'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'

Test the value 'inherit'
PASS getComputedStyle(parentElement, '').getPropertyValue('will-change') is 'contents, opacity, top'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'contents, opacity, top'

Test that will-change is not inherited by default
PASS getComputedStyle(parentElement, '').getPropertyValue('will-change') is 'opacity'
PASS getComputedStyle(element, '').getPropertyValue('will-change') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
